Modified the node shape transfomer so that NUMERICAL answers will map size to a continuous range.
If an Alter question is defined with a NUMERICAL answer, and Literal numbers are set for the min/max bounds on that question, the node size for a given answer will be determined as:

[(answer - minAnswer) / (maxAnswer - minAnswer)] * (maxScale - minScale) + minScale

I've set maxScale and minScale to 4.0 and 1.0, respectively. So, the largest possible node is 4x the height/width of the smallest. These parameters could be added to the question-editing page later so they could be modified by the interview author, but this should suffice as an initial implementation.