Added more rendering options for all node/edge transformers.
The transformers for node shape/color/size and edge color/width all looked for "0", "1", and "2" in the associated answer and were drawn accordingly. We will commonly have questions with more than 3 selection options, so this commit adds colors/shapes/sizes for all five transformers. They will now look for "3" and "4" in the answer as well, providing five rendering options for responses to a selection question.