moved graph node labels from centered to the bottom-right of the node in NetworkService added functions to create/retrieve graph layouts from outside the NetworkService/NetworkImage classes, and to set the layout used for graph rendering
interviewingnetworkpage:
changed default size for network nodes to 1.75x what it was
changed response-driven node sizes for graphs to be slightly larger as well
changed text sizes to be slightly larger

graphs drawn for the same interview/expression/layout-type are now drawn using the same layout. So, if multiple questions are shown with the same adjacency expression (or if one question is shown twice) the layout won't change between viewings.

Signed-off-by: ethand <ethan.drucker2@gmail.com>